added rate limiter.

+++ src/main/java/com/project/library/controller/AdminController.java
@@ -2,6 +2,7 @@ package com.project.library.controller;
 
 import com.project.library.dto.BookDTO;
 import com.project.library.service.AdminService;
+import com.project.library.service.RateLimiterService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
@@ -24,8 +25,13 @@ public class AdminController {
 
     private final AdminService adminService;
 
+    private final RateLimiterService rateLimiterService;
+
     @PostMapping("add-book")
     public ResponseEntity<?> addBook(@RequestBody BookDTO bookDTO) {
+        if (!rateLimiterService.tryConsume()) {
+            return ResponseEntity.status(HttpStatus.TOO_MANY_REQUESTS).body("Too many requests. Please try again later.");
+        }
         try {
             return new ResponseEntity<>(adminService.addBook(bookDTO), HttpStatus.OK);
         } catch (Exception e) {
@@ -36,6 +42,9 @@ public class AdminController {
 
     @PutMapping("update-book/{id}")
     public ResponseEntity<?> updateBook(@PathVariable Long id, @RequestBody BookDTO bookDTO) {
+        if (!rateLimiterService.tryConsume()) {
+            return ResponseEntity.status(HttpStatus.TOO_MANY_REQUESTS).body("Too many requests. Please try again later.");
+        }
         try {
             return new ResponseEntity<>(adminService.updateBook(id, bookDTO), HttpStatus.OK);
         } catch (Exception e) {
@@ -46,6 +55,9 @@ public class AdminController {
 
     @DeleteMapping("remove-book/{id}")
     private ResponseEntity<?> removeBook(@PathVariable Long id) {
+        if (!rateLimiterService.tryConsume()) {
+            return ResponseEntity.status(HttpStatus.TOO_MANY_REQUESTS).body("Too many requests. Please try again later.");
+        }
         try {
             adminService.removeBook(id);
             return new ResponseEntity<>(HttpStatus.OK);
